# Optimize global reduce_agg aggregation