… velox/functions/prestosql/aggregates/ReduceAgg.cpp