Summary:
Optimize global reduce_agg to evaluate lambda expressions log2(n) times where n is number of rows. The original implementation evaluates expressions n times.

Add benchmark. Initially, reduce_agg is 60x slower than sum. After optimization it is "only" 4x slower.

Before:

```
============================================================================
[...]l/aggregates/benchmarks/ReduceAgg.cpp     relative  time/iter   iters/s
============================================================================
reduce_agg                                                   1.14s   873.85m
sum                                             67438.%     1.70ms    589.31
```

After:

```
============================================================================
[...]l/aggregates/benchmarks/ReduceAgg.cpp     relative  time/iter   iters/s
============================================================================
reduce_agg                                                  6.16ms    162.39
sum                                             360.26%     1.71ms    585.03
```

Differential Revision: D49192172